Add stub functions for libatomic calls used by the new synchronization code. The libatomic calls are generated in the 32-bit runtime because we compile it using LLVM's PNaCL backend (le32) and this forces generation of libatomic calls for atomic instructions. So by the time we get to compiling the runtime LLVM IR for an actual target, the atomic instructions have already been lost. These stubs are inline routines that convert the calls back into atomic instructions so the actual target backend can lower them appropriately. If the final backend requires libatomic, this should result in actual calls to the libatomic routines and require the end user to add -latomic to their link line. (Just like they have to add e.g. -lpthread.)
(Performance on a target that uses libatomic may or may not be
acceptable. If we have one we care about, we should measure how well
it works. I'm not sure there are actually any that are supported.)

This is rather a hack to workaround a n edge case issue with our very
peculiar runtime compilation approach. It may however be the simplest
most reliable approach. The disadvantages with this are complexity and
that it turns all compare-and-swap operations from weak ones to strong
ones. (A weak CAS may fail supriously requiring a retry loop.
Theoretically it can be more performant. A strong one only fails if
the compare fails.) I doubt this is a significant issue but it is a
consequence of this workaround.