Disable Warning C4800 in MSVC
This is the infamous "forcing value to bool 'true' or 'false' (performance warning)" oddness.

This warning is no longer emitted by VC2017, so by leaving it in, we're getting warnings (-as-errors) that only repeat in VC2015. Even in VC2015, it's basically a giant pile of low-information noise that only causes weird Windows-specific breakages.

See https://developercommunity.visualstudio.com/content/problem/78259/visual-studio-2017-no-longer-generates-a-c4800-war.html for more info.