Fixes to attempt getting all tests to pass.
This removes ParamMap::empty in exchange for a breaking change requiring writing
"ParamMap{...}" when passing a ParamMap as a literal in the function call slot. There
are probably other solutions here, but I felt this is worth considering as the static
global seems like a bad idea. (In general, I think requiring the type name at the call
site for various things might make our API easier to use and code easier to read, but
this needs more thought.)

Another approach for RealizationArg would be to use inline storage for
Realization and add a move constructor. This would save the new for
the case where it gets a literal braced list. That design likely has
to do one set of refcount increases however. (Because a Realization
may be long lived.) If a Realization is constructed to be passed to
the RealizationArg, then ref counts were likely increased anyway so
the inline storage approach may be simpler.