Add a way to add arbitrary data to Funcs when tracing
This adds a new trace event that allows tagging Funcs with arbitrary text data, which is streamed to the trace listener when(if) the Func is realized in tracing mode.

Although this could be used for any trace-related purpose, the long-term purpose I have for it is to upgrade HalideTraceViz:

(1) in the short term, allow using the existing flag vocabulary for layout, but via adding the tags to the Halide source code, rather than requiring a separate set of flags that are maintained in a shell script. (I actually such a change implemented locally, but left it out of this PR to simplify the review; I can add it here or as a separate, dependent PR if you like.)

(2) In the longer term, to augment the flag vocabulary for HalideTraceViz in order to allow a reasonable auto-layout facility, so that visualizations of arbitrary pipelines can be done with little-or-no manual flagging; this would be especially useful for use with the Python bindings in environments like Jupyter, Colab, etc.

All tag data is (deliberately) emitted at the beginning of a pipeline; the intent here is to allow the tracing tool to be able to (eventually) make layout decisions at the start of rendering (e.g. with knowledge about intermediate realizations that it won't encounter until later), so that the current streaming architecture can be maintained.

Note that (as written) this adds one byte to every trace packet; this could probably be eliminated, but seems a reasonable cost.

I'm not wild about the nomenclature of "trace_tag", and am open for better suggestions.