Remove UB from add_custom_lowering_pass
Casting the deleter function to a different function type, then calling it, produces UB warnings from UBSan. Rework to use a no-arg std::function for the deleter, and wrap up the necessary deletion in a lambda to avoid the casting.

(Why aren't the custom lowering passes just passed in via unique_ptr or shared_ptr?)