Encapsulate some logic in work queue methods.
Assert work queue is zeroed at init time instead of forcing it to be
zero. All zeros is the default static init state and indicates no
threads are waiting on condition variables as the fast_synchronization
logic guarantees a zero value is the quiescent state for a mutex or
condvar.

Zero work queue at shutdown time to return it to proper
initializations state. (And to reflect its actual state after threads
are shutdown.)