Add Target::LegacyWrappers feature
This disables the default generation of the classic 'buffer_t' wrapper code for AOT compilation; the code is still present, but must be deliberately opted in by using the new Target::LegacyWrappers flag when compiling. If you need to enable this flag for your code, you should strongly consider upgrading it to use halide_buffer_t instead (perhaps via the Halide::Runtime::Buffer<> class), as support for the classic 'buffer_t' struct will likely be removed at some point in the future.