Split NO_INLINE in two
The current NO_INLINE macro actually means "don't inline if we are in user code, but if we are compiling Halide, do whatever".

This PR adds a new HALIDE_NEVER_INLINE macro, which is the opposite of HALIDE_ALWAYS_INLINE: it never inlines code, period. (It is meant to be useful in header-only libraries like halide_image_io.h, where being able to specify some functions as never-inline can be a useful hint to the compiler.)

The existing NO_INLINE macro is renamed to HALIDE_NO_USER_CODE_INLINE to more accurately reflect its meaning (and to give a HALIDE_ prefix to it, which it should have always had).