Use add_trace_tag() to allow sending layout info to HalideTraceViz
This is essentially a revised version of #2838, but instead of attempting to replicate the flags 'language', we instead expose the the Config struct (via a shared header), and serialize/deserialize it into a trace_tag that HTV can interpret.

As proof-of-concept, it converts apps/bilateral_grid and apps/local_laplacian to use trace_tags instead of command-line args. (We may or may not want those changes to land as part of this PR.)

(Note that this doesn't provide a way to set global flags such as `--timestamp`; those could be added in the future via another struct.)