Bounds calculations needs to understand strict_float
Bounds(Call) assumes the worst for calls it doesn't know about, so strict_float(foo) becomes (neg_inf, pos_inf) rather than (foo, foo) (a la 'likely'), exploding bounds checking.

Note 1: this needs a test to be written; I haven't done so yet because I'm puzzled as to how this isn't covered in our existing tests (it was found inside Google code).

Note 2: it's possible that some of the other 'likely' special cases need special-casing here too; I'm not familiar enough with the Bounds code to have decided whether that's the case or not. (The fix "as-is" here fixes the cases I know about.)