Convert camera_pipe to use HalideTraceConfig struct
Done mainly as a proof-of-concept to further flush out work needing to be done; not sure if we should land this, because if we do, we'll have no remaining usage of most of the flags args (aside from --timestamp, --size and a few others), and they will inevitably bitrot. (OTOH, maybe we're better off doing this conversion and removing the flags support? Opinions welcome.)