Add trace_tag for 'global' visualization params
This allows us to set (e.g.) timestep or framesize via trace_tags in the code, rather than command-line flags.

Note that this is of limited use at present, as the existing flags tend to be sensitive to input data and thus may be better specified on the command line; however, this provides a good mechanism to specify future params related to automatic layout.