Rework FuncConfig to allow merging/optional fields
The goal here is to allow some-but-not-all fields in FuncConfig to be specified, so that we can override some-but-not-all aspects of a visualization.

As it is, this means you can specify a FuncConfig via a trace-tag, then tweak/override values via --flag values without having to rebuild/rerun the Halide generation code.

This will become especially useful after auto-layout for visualizers is put into place, as that will allow you to just use auto-layout to get a rough visualization, and then (if desired) tweak it piece-by-piece (by adding trace-tag and/or --flags) to get a pleasing layout.

Note that I implemented the structs by using 'magic invalid values'; a perhaps-cleaner-but-more-expensive approach would be to have each field allocated by std::unique_ptr, with nullptr meaning "not set"; I rejected that because it makes setting each field considerably more awkward.

(To the Googlers watching: yes, mergeable-structs-with-optional-fields sounds a lot like protobufs; no, we're not bringing in protobufs just to accomplish this.)