Remove threading from compile_multitarget
This helped build throughput a little bit, but can introduce a subtle nondeterminism in multitarget builds: different unique_names can be chosen if the modules are built in different order, since the unique_name_counters global is never reset; build tooling that looks for nondeterminism in builds will (rightly) occasionally find it here. Since the performance improvement from this threading was modest in the first place, just remove it entirely.