Fix race condition in tracing runtime
If multiple threads requested packets at the end of the buffer, you could end up with a bad packet if the back-out was ordered in a certain way. Instead, don't try to back out, just let all requests fail till the next flush(), but record the total overage so we know how much to ignore.