Change MAX_THREADS back to 64
It was bumped to 256 by the fast-sync changes, but since we still pre-emptively fill up a threadpool with number-of-available cores, this can cause a jump in threads allocated on some higher-end systems, leading to pointless memory pressure. Let's move the limit back to where it was for now; we can bump it back up after we're confident we have appropriate checks in place everywhere donwstream.