Add support for ThreadSanitizer
Add Target::TSAN and support for it. Unlike the previous MSAN support, though, this actually just adds the ThreadSanitizer LLVM pass at the right spot, meaning that (in theory) both Halide-generated code *and* our runtime code should be TSAN-checked.

Still experimenting, but if this pans out, it should allow us to replace the previous MSAN hackery (mark buffers as touched) with "real" MSAN checking (which will be many times slower of course, but that's OK).

Also would allow adding ASan (and maybe UBSan, though it's not clear if C++ undefined behavior warnings are useful or interesting to us...)