VectorizeLoops needs to specialize Call::trace even more
begin/end_realization events got vectorized in a way that increased the number of dimensions by the vector width in a bizarre way, e.g. if the realization was sum(vectorize(4)), a 1-dimensional realization would emit a 4d begin_realization, with the data in the form (min0..3, extent...3). Code downstream (e.g. HTV) could probably recognize and deal with this situation but it seems better to try to handle it here.