Add an auto-layout mode to HalideTraceViz
The general idea is that
(1) Tracing.cpp now adds trace-tags that contain the type info and dimensions/extent of all Funcs touched (dimensions are a conservative guess based on calling boxes_touched())
(2) HTV now has an `--auto_layout` flag that, when set, attempts to lay out all touched funcs in a way that isn't crazy: in the order touched, in a set of left-to-right/top-to-bottom squares, roughly maximizing space used, zooming as it sees fit, adding labels, etc.
(3) By default, any trace-tags that are specified in Halide code (or on the command line) override the auto-layout flag; this is useful in that you can use it to start with auto-layout, then gradually tweak it to a pleasing layout by adding tags (e.g. if you want to use it for a presentation). OTOH, if you use `--auto_layout` on a pipeline that already is tweaked, you can get bizarre results (cf camera_pipe for an example); you can pass the flag `--ignore_tags` to have auto layout ignore any custom tags and just use pure auto layout.
(4) Added a `viz_auto` shell and makefile usage to a handful of apps; that said, I think the better longterm usage for autolayout is more likely to be by breaking HTV up into library form and integrating it into the RunGen utility.
(5) Doesn't deal with Tuple-valued Funcs well (known existing limitation of HTV); hasn't been tested with define_extern() yet so might be broken there.

Aside from commenting on the code, feel free to pull and try it on various pipelines; feedback on improving the layout is welcome. (Just keep in mind that this autolayout algo isn't intended to be aesthetically pleasing; it's intended soley to quickly lay out things in a way that makes visualization of code patterns comprehensible to a typical Halide coder. If we can also make it pretty, that's nice too, of course.)