Fix benign race in halide_upgrade_buffer_t
If you have a define_extern that uses the legacy_buffer_t wrappers, and that extern is inside a parallel loop, there is a benign race condition: the input buffer(s) are always 'upgraded' at the end (in case the call is a bounds query). This causes a benign data race in the non-bounds-query case since another thread can be reading from the input buffer at the same time. The fix: when generating the wrapper for define_extern, only upgrade input buffers if it is known to be a bounds query.