Disallow unsafe Buffer::crop calls()
Previously we explicitly allowed crop() calls that expanded the Buffer, as the implementation of Func::infer_input_bounds() relied upon it. This is a dangerous and bad idea. Add assertion-checks that require crop() goes inward, and fix the implementation of Func::infer_input_bounds() to not rely upon the previous behavior.