Preserve const-ness when casting pointers in HalideTraceViz.cpp
I get build errors on lines 820 and 821 of HalideTraceViz.cpp, pertaining to the loss of `const` qualifiers during the casts on the relevant lines. This change preserves the const-ness from the function’s parameters – making these pointers `const` should not have any effect on the functions’ operation as they are only read from and not written to.