Handle more complex instances of select with indeterminate on one side
Consider: select(x > 0, a/x, select(x < 0, b/x, indeterminate()))

The top-level select evaluates both sides, if x is positive, the RHS
then throws an error.

This commit changes evaluation of an indeterminate() to UB, which sucks,
but means we can still evaluate both sides of a select. The other option
would be to only evaluate both sides of a select if neither side could
throw an error. This option generates simpler IR though in
non-pathological cases.