Implement a new model of thread reserve counting for async.
Writing down the design idea here in lieu of hopefully forthcoming design doc:

Any Parallel For or Fork that contains one or more Acquire nodes will
*itself* occupy a thread. These Acquire nodes can be directly
underneath the Parallel For or Fork, or inside a child Parallel For or Fork
node. If the Parallel For or Fork node only contains direct Acquire nodes, it
consumes one thread. If it contains no Acquire nodes at all, it
consumes zero threads.

In a Block, the number of threads consumed is the maximum of all its
Stmts.

A Fork node takes the sum of all its branches because in the worst
case, each branch could be executed on a different thread. In addition
the Fork consumes zero or one *additional* thread by rule 1. Note that
a Fork containing no Fork or Parallel For nodes always takes zero or
one thread.

A Parallel For consumes as many threads as its closure and the number
of iterations that are invoked. There is a question here on how many
threads to reserve for the Parallel For. The initial design is to
account for it as one job's worth of threads and to allow multiple
iterations against the top-level work queue only. Parallel Fors that
consume zero threads (no Acquire nodes) operate as they always have:
without constraint.

A job reserves its thread count from a pool. For top-level jobs, this
pool is on the work_queue. For subjobs, it comes from their parent
pool, which is initialized to the number of threads the parent
reserved from its parent (or the work_queue) -- that is the reserve
count for the parent. (There is one potential issue in the current
implementation that a top-level halide_do_parallel_tasks does not have
a min_threads count for the Fork itself. Thus theone thread that gets
added there may not be subtracted from the work queue total in the
reservation.)