Fix failing tests.
The following tests were failing on HVX on master after interleave-bug
was merged into master.

    correctness_argmax
    correctness_bool_compute_root_vectorize
    correctness_fuse
    correctness_in_place
    correctness_multi_way_select
    correctness_simd_op_check
    correctness_undef

Fixed these tests now. The problem with all, except simd_op_check was
that we weren't handling scalar loads and stores in HexagonAlignment.h
We do, to a certain extent only though. As in we look for "too much" alignment
but we do not crash any more like before.

simd_op_check was failing all valign and vlalign issues and that was due to a bug
in AlignLoads that could be traced to refactoring the alignment logic into
HexagonAlignment.h