Add optional wrappers (V2) to AOT headers to take Halide::Runtime::Buffers as arguments
This is a reworked variant of #3206 that dispenses with the Google-style convention and requires all Buffer arguments to be mutable, whether by ref or ptr. Note also that (unlike the previous patch) it disables the implicit `operator halide_buffer_t*` provided by the `Buffer` class, in an effort to ensure that it's harder to pass a wrong-static-type Buffer to a given function (and, of course, with the idea that implicit conversions are always preferable to avoid when not absolutely necessary).

Not sure if this is ready or desirable to land, but putting it out there for discussion purposes.