Buffer<const T>::copy_from() should be illegal
It should not be legal to call copy_from() on a Buffer with a const value type, but (thanks to explicit casting in the implementation which subverts the usual checks), it is. Added a static_cast() to prevent this.

Also, a drive-by fix to make is_bounds_query() a const method.