Renable full runtime folding validity check for dynamic footprint case.
The above could fail due to a race condition in reseting the
boundaries of fold (head/tail) at the end of a loop. This was fixed by
adding "_next" versions of the head and tail counters which allow
setting the value to update the corresponding counter to for the next
loop *after* its semaphore has beein acquired.

Fix a bug where the dynamic footprint name could get an extra ".head"
appended as that should only happen if the Func is async.