Tweak simplification to ensure something compute_at a guardwithif on stack
GuardWithIf is a really important TailStrategy, because it's the only
safe way to split an update definition that reads from an input image
that isn't a multiple of the split factor. However things that are
compute_at a GuardWithIf split are not landing on the stack, crippling
performance and making this tail strategy unusable :(

This PR makes two changes to simplification to handle the IR patterns
produced by guard with if and help bound_small_allocations understand
that these things are bounded and can go on stack.

1) Substitutes in the RHS of things like min(..., var) and max(..., var)
as we currently also do for add, sub, and mul.

2) Makes unify_duplicate_lets mutually recursive with the simplifier. It
opens new simplification opportunities, which in turns opens more
deduplication opportunities, so unless we want to iteratively run the
passes until convergence they should be mutually recursive.

Added a test to ensure that something computed_at a GuardWithIf split
does indeed go on the stack.

Also made a drive-by change to reduce the verbosity of the debug stmts
in Pipeline.cpp. I'm sick of seeing thousands of them when benchmarking
things.