Fix profiling when a stage is offloaded to HVX.
The HVX device side runtime maintains its own profiling
state that is sampled by the sampling thread. However, if a stage is
offloaded to HVX, the current_func is set on the host, but the HVX
side runtime won't know of it until it executes a halide_profiler_set_current_func
call on its own. This means all the profiling information up until such a point
is lost and counted as "overhead"