Expand LLVM_CONFIG and CLANG for test_apps
The test_apps target recursively calls sub-makefiles (which can change the WD), but the buildbots define LLVM_CONFIG and CLANG as relative paths, so these can become invalid when building apps. (Why this hasn't caused more failures in the past is a bit of a mystery.) Use $(realpath) to explicitly expand these vars for the sub-makefile calls to address this.