Revise apps/ to depend only on distrib/ folder
Up to now, the apps/ folders all got the privelege to look inside Halide; this revises them to depend only on the contents of the distrib/ folder, as should be the case for all 'normal' Halide-using apps. This patch mostly mimics the approach used by python_bindings/, which already has this assumption. Details include:
- Makefile.inc references HALIDE_DISTRIB_PATH rather than HALIDE_SRC_PATH/HALIDE_BIN_PATH
- having the test_apps target define HALIDE_DISTRIB_PATH instead of HALIDE_SRC_PATH/HALIDE_BIN_PATH
- moving the 'bin' dirs to be inside the apps/ subfolders (so that they are invariant whether built by test_apps or individually and making clean address them too
- Fixing some assumptions made in the matlab code
- adding tools/halide_malloc_trace.h to the distrib folder

This also adds a `halide_config.make` stub to the distrib folder (as we have already been doing for CMake and Bazel), so that we can hopefully remove the use of LLVM_CONFIG in the apps/ support as well. (I'm deliberately avoiding doing that in this PR in order to ensure that just the distrib-only change doesn't break anything on its own.)

@dsharletg -- I haven't (yet) tested the Matlab-specific changes; they probably could use a closer look from someone more familiar with that code (ie, you)