Refactor RunGen
Move most of the interesting parts of RunGen.cpp into a (new) RunGen.h file; rename RunGen.cpp -> RunGenMain.cpp, since it now contains mostly just a main() function and a little support code.

The motivation here is to be able to re-use this code for other tests and benchmarks downstream.