Only emit default-values to metadata when defined
Our metadata format was designed to emit null fields for def/min/max of scalar values when said values aren't defined; in the case of Generators with Input<SomeScalar>, we were emitting a default of zero (rather than "no default") for the case of no-default-specified. This fixes the issue and updates the test.