Optional constraints for Input<Func> and Output<Func>
This is a sketch of a possible way to address Issue #3306: `Input<Func>` and `Output<Func>` have a new method, `dimensions_and_alignment()`, which returns a struct that lets you get/set the min/extent/stride/host-alignment; these values are entirely ignored if the given Input/Output isn't AOT-compiled as a Buffer.

This works fine and I feel it's robust, but I'm not sure if this API is one I'm happy to support forever. That said, it would make it a lot easier to make composable Generators per the details in the issue.