Remove support for ImageParam/Param<> in Generators
`Input<>` has been the preferred way to specify inputs to Generators for a long time now; this PR proposes to finally remove the support for using `ImageParam` and `Param<>` entirely. (This allows us to simplify various bits of code inside Generator and Parameter, with the `is_explicit_name` nonsense a welcome goner.)

Note that I'm deliberately leaving `RegisteredParameter` for now, as it will let us continue to detect (and assert) if someone mistakenly attempts to use `ImageParam`/`Param<>` in a Generator, rather than just quietly misbehaving. (I'd like to remove that as well at a later date.)

(Note that this does NOT remove support for the `build()` method; it's still allowed to use `build()` instead of `generate()` and `Output<>`)

The gating factor here is whether all downstream users are ready to accept this change; we may need to delay landing this depending on that. (Google is fine with this change, as all of code has been migrated to use `Input<>` for quite a while now.)