Avoid (some) UB in ModuleRemainder
In some cases, you can end up with expressions like `(0 - -2147483648)` for the remainder calculated for a Sub node; if running under a sanitizer like Clang's UBSan, this will be flagged as a fatal error. To avoid, check for overflow and revert to the (0, 1) case for the Add, Sub, and Mul nodes.

(Unfortunately, the test cases that reported this are immensely complex, and I haven't boiled down a simple test case to add here yet.)