Add Buffer::as_ptr() and as_ptr_const()
These are analogous to as() and as_const(), but return pointers rather than refs; this is useful for coding in C++ styles that require pointers for mutable references, when you have (eg)

    void DoSomethingWith(Buffer<const int> *input);

     Buffer<int> foo;
     DoSomethingWith(&foo);  // won't compile
     DoSomethingWith(foo.as_ptr<const int>());  // OK
     DoSomethingWith(foo.as_ptr_const());  // also OK

Note that I deliberately avoid as_const_ptr() as the name, as that would imply to me the pointer returned is always const, which isn't the case.

Also note that I'm deliberately avoiding rvalue overloads for this, as it's not clear to me that it's an interesting use case.