Add runtime wrapper for abort()
Windows Debug runtimes stall the `abort()` call by presenting a modal dialog. Add a runtime overload for windows that does the equivalent raise/exit combination as documented by MSFT instead and use it instead of raw abort(). Ditto for the compile-time Error.cpp in the compiler. (Note: this makes no attempt to convert apps or tests that explicitly call abort; this is just for the Halide compiler and runtime itself.)