Add classic auto-schedule testing to apps/resnet
Note that (unlike most other apps) this doesn't attempt to run all three in the same run; it selects one of the options via command-line flag. 'make test' runs all three in sequence.

Note also that the 'manual' schedule is known to be awful (it's just enough compute-roots added to make the runtime tractable); improvement is left as an exercise to the reader.