Fix possible n^2 arg evaluation in Bounds::visit(Call)
We could evaluate the args redundantly, due to the eager calculation of `new_args` in case it's a pure function call; for a deeply-nested Expr, this could end up with a meaningfully n^2 delay. We never noticed this before because the main candidates to be deeply nested were bitwise ops (which were previously eagerly simplified). Fortunately, fix is simple: rearrange the `new_args` calculation to be done only if we know the other criteria for a possible PureExtern/Image function are satisfied.