Fix require() for HVX backend
HVX (and possible OpenCL) have code to eliminate bool vectors by converting them into a mask; this breaks scalarizing require() if the condition is a vector, as it assumes the condition is still bool, vector or not. This just inserts a workaround to re-boolify the vector when scalarization is necessary.