Refactor HL_LLVM_ARGS parsing
Currently, we parse the HL_LLVM_ARGS environment variable exactly once (when we init LLVM). This proposes to refactor that to allow resetting the LLVM flags once per compile:

-- defines CodeGen_LLVM::set_llvm_command_line_options() as a virtual method; the default implementation resets all LLVM flags to their default values, then re-parses from HL_LLVM_ARGS as before

-- subclasses are allowed to override this to make additional settings; the existing usage in Codgen_Hexagon uses this approach now (instead of overriding `compile()`, and an override for Codegen_X86 was added to set `enable-mssa-loop-dependency=1`, which can improve compile time substantially in some cases (see @alinas for details)

That said: I'm not 100% satisfied with this approach, because it means we're relying on mutating LLVM flags (aka "global variables") to get the settings we want; if the Halide compiler is being used from multiple threads in the same process, you can guess what sorts of shenanigans could potentially ensue. (That said, the risk is small, because in practice, we ~never use the same process to generate different architectures in multiple threads... at least, so far.)

On the other hand, if we need to set things in LLVM that are only controllable via command-line flags, our options would seem to be:
(1) take this approach and hope things work
(2) restrict LLVM usage to a single thread at a time in Halide
(3) patch LLVM to allow the settings to be controlled per-LLVMContext

This PR is of course approach 1; it's not optimal, but it may be acceptable for the short term. Opinions welcome.