Set llvm::TargetOptions::UseInitArray = true
This causes llvm to put ctors into the `.init_array` section when possible (rather than the .ctors` section.

This can dramatically improve performance for some targets (esp aarch64) when the `lld` linker is used: older linkers (eg, gold) tend to do legacy conversion of .ctors to .init_array, but lld does not.

(It's not clear to me why we default this to false -- historical reasons? mistake? ignorance? -- or if there are any downsides to making this change.)