Revert recent bitwise ops changes
Revert the changes from #3847, #3867, and several other PRs which addressed followup issues. The PRs in question changed the behavior of the bitwise ops (& | ^ << >>), and while these changes are arguably desirable in the long run, they change the language definition in ways that will require downstream code to carefully check for the new behavior (the shift operators in particular have this issue).

The proposal here is to revert this for now, with the plan of reintroducing these changes in subsequent PRs, but in a more gradual and staged way that will allow downstream users to update their code for the new behavior more carefully.

(Note to reviewers: this is a combination of spot-reverts of many different commits; I think this is back in the correct 'pre' state but a careful review is appropriate.)