Add GENERATOR_OUTPUTS var to apps/ Makefiles
This gives a convenient bottleneck to customize the Generator output for (most) of the apps/ built via Makefiles; default to also generating .stmt and .s files, so that people examining apps/ output for benchmarking, learning, etc can see this output without having to do surgery in the Makefile.

(Note that some of the apps/ with Makefiles that are already heavily custom, e.g. the hexagon benchmarks, weren't touched by this. It would be nice to regularlize all the apps build files at some point in the future.)