define CLANG relative to LLVM_CONFIG
For ~ever, we've required Makefile users to define CLANG as well as LLVM_CONFIG as env vars; in practice, CLANG should always point to the the one that could be deduced by LLVM_CONFIG. This changes the CLANG default value to that one (instead of 'system clang') and removes the recommendation that it be set. (You can still override CLANG if you really want to.)

If this is accepted, the buildbots will be subsequently updated to stop setting CLANG.

Note that I skipped tests/opengl/vagrant since it seems to be deliberately using ancient LLVM versions that we don't support anyway -- is it stale?

(Note that the CMake build rules were already doing it this way)