Give apps/autoscheduler its own logging function
This replaces the use of `debug()` in apps/autoscheduler code with its own logger (named `aslog()` for now, short for auto-scheduler-log.)

The motivation here is to allow supressing of the autoscheduler log output without also having to supress normal codegen output.

aslog() has the same semantics as debug(), except that it preferentially tries to read from HL_DEBUG_AUTOSCHEDULE for a verbosity; if that isn't set, it falls back to the same value as HL_DEBUG_CODEGEN. (Thus, the autoschedule output can be suppressed by HL_DEBUG_AUTOSCHEDULE=-1.)

Note also that all use of std::cerr that was for anything other than a fatal error was converted to aslog(0).

And yes: "aslog" isn't a very good name; I'm happy to replace it with literally anything you'd like before submitting this :-)