Refactor simd_op_check.
- Refactor to the target agnostic aspects of the test into a base
class in a common header file
- Use the base class to derive two classes - one for HVX specific tests
and the other for all the other targets that simd_op_check used to
test.

This change is done because HVX departs from all other architectures
tested by simd_op_check
  - It isn't a 'host' arch so to test this now, HL_TARGET will have to
be such that host is hexagon. This is fine, however, it breaks the way our
buildbot is configured.
  - It requires some changes that other architectures do not, such as
setting the min in dim.0 of all image params to 0 or a value aligned to
a  128 byte boundary.

These are relatively small changes that may not, on the surface, warrant
such a big refactoring but it is better in the long run, IMHO.