Convert autoscheduler weights to a single file
The main gist of this change is to put all of the weights used by the autoscheduler into a single file (instead of a directory of multiple files); the goal is to make saving 'overnight' tuned weights for a pipeline simpler, via having only a single `.weights` file rather than a directory of things.

There are various drive-by fixes, of course:
- The 'Weights' struct moved into its own file, and now manages its own loading/saving, in both the classic dir-of-files format, and new, single-file format. (The single-file format is very simpleminded but likely more than adequate; it has a trivial signature at the front so some smart upgrading over time should be easy to do.)
- DefaultCostModel.cpp got its own .h file, instead of being incestuously connected to the ABC it implements
- existing `weights` folder was deleted; the equivalent data is now in `baseline.weights`
- added the `weightsdir_to_weightsfile` utility for any downstream consumers that have custom weights in the 'old' format that they need to update to the 'new' format.
- Removed some unused code here and there.

Note that HL_WIEGHTS_DIR (and HL_WEIGHTS_OUT_DIR) are still used, but are 'smart' for now; if you pass a pathname that ends in `.weights` it loads as a single file, otherwise it assumes an old-style directory. (I didn't try to smarten the env var names because we really need to eradicate the use of env vars for this purpose anyway. but that's a job for another day.)